pyproject.toml
[tool.poetry]
name = "dia-mvc"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.4"
django-import-export = "^2.5.0"
pandas = "^1.2.1"
sklearn = "^0.0"
django-forms-bootstrap = "^3.1.0"
django-sendsms = "^0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
